[#2127] Refactor
